requirements.txt
APScheduler==3.0.2
feedparser==5.1.3
futures==2.2.0
oauthlib==0.7.2
pycparser==2.10
pymongo==3.0.3
pytz==2015.2
PyYAML==3.11
requests==2.5.3
requests-oauthlib==0.4.2
six==1.9.0
tweepy==3.3.0
tzlocal==1.1.2
